refactor(layout): restructure component organization and update exports
Moved Layout and LayoutHeader components into a nested "components" directory for better organization. Updated import paths accordingly to reflect the new structure. Removed the now-unnecessary index file for LayoutHeader, simplifying the overall export statements. This change enhances clarity in component hierarchy and maintainability of the codebase.
