--- src/components/LayoutHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './LayoutHeader'
