--- src/components/LoadingIndicator/LoadingIndicator.module.css.d.ts
@@ -1,5 +1,0 @@
-declare const styles: {
-  readonly "container": string;
-};
-export = styles;
-
